refactor(scripts): move kadence-packages helper to scripts/helpers
